requirements.txt
pyrogram>=2.0.0
wikipedia==1.4.0
tgcrypto>=1.2.5
python-dotenv>=1.0.0 
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
APIFlask==2.1.1
apispec==6.6.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
environs==11.0.0
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
Flask-Admin==1.6.1
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
flask-mongoengine-3==1.0.9
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanize==4.9.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-chroma>=0.1.4,<0.2
langchain-openai>=0.2,<0.3
langchain-text-splitters>=0.3,<0.4
langchain-experimental>=0.3,<0.4
langgraph>=0.2.20,<0.3
langsmith>=0.1.125,<0.2.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-mongoengine==0.31.2
mongoengine==0.28.2
multidict==6.0.5
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
regex==2024.5.15
requests==2.32.3
sentry-sdk==2.9.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.4.2
tiktoken==0.7.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uWSGI==2.0.26
vine==5.1.0
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.3
WTForms==3.1.2
yarl==1.9.4
pysqlite3-binary==0.5.3
tabulate==0.9.0